apps/api/requirements.txt
# Web Framework
fastapi==0.109.0          # Modern Python web framework with automatic API docs
uvicorn==0.27.0           # ASGI server to run FastAPI
pydantic==2.5.0           # Data validation using Python type hints
pydantic-settings==2.1.0  # Manage environment variables

# HTTP Requests
httpx==0.26.0             # Modern HTTP client for calling Oura API

# Database
sqlalchemy==2.0.25        # ORM for database operations
psycopg2-binary==2.9.9    # PostgreSQL adapter
alembic==1.13.1           # Database migrations

# AI
anthropic==0.18.1         # Claude API client

# Background Tasks
celery==5.3.6             # Distributed task queue
redis==5.0.1              # Redis client for caching and Celery broker

# Utilities
python-dotenv==1.0.0      # Load environment variables from .env file
